+++ ot_simple_rest/handlers/eva/db_connector.py
@@ -1,3 +1,5 @@
+from typing import List, Dict, Any, Union
+
 from tools.pg_connector import PGConnector
 from copy import deepcopy
 
@@ -9,6 +11,10 @@ __version__ = "0.1.2"
 __maintainer__ = "Andrey Starchenkov"
 __email__ = "akhromov@ot.ru"
 __status__ = "Production"
+
+
+class QueryError(Exception):
+    pass
 
 
 def flat_to_set(arr):
@@ -641,6 +647,17 @@ class PostgresConnector(PGConnector):
         dash_id = self.execute_query("SELECT id FROM dash WHERE name = %s;", params=(dash_name,))
         return dash_id
 
+    def _get_dash_groups(self, dash_id: int) -> List[Dict[str, Any]]:
+        return self.execute_query(
+            """
+            SELECT "group".name, "order" FROM dash_group
+            JOIN dash ON dash_group.dash_id=dash.id
+            JOIN "group" ON dash_group.group_id="group".id
+            WHERE dash_id=%s;
+            """,
+            params=(dash_id,), fetchall=True, as_obj=True
+        )
+
     def get_dashs_data(self, *, group_id=None, names_only=False):
         if group_id:
             dashs = self.execute_query("SELECT id, name, body, round(extract(epoch from modified)) as modified "
@@ -654,10 +671,7 @@ class PostgresConnector(PGConnector):
             dashs = [d['name'] for d in dashs]
         else:
             for dash in dashs:
-                groups = self.execute_query('SELECT name FROM "group" WHERE id IN '
-                                            '(SELECT group_id FROM dash_group WHERE dash_id = %s);',
-                                            params=(dash.id,), fetchall=True, as_obj=True)
-                groups = list({v['name']: v for v in groups}.values())
+                groups = self._get_dash_groups(dash.id)
                 dash['groups'] = groups
         return dashs
 
@@ -687,6 +701,20 @@ class PostgresConnector(PGConnector):
         dash_data['groups'] = groups
         return dash_data
 
+    def _add_dash_to_group(self, dash_id: int, group: Union[str, dict], conn):
+        if isinstance(group, str):
+            self.execute_query(
+                'INSERT INTO dash_group (group_id, dash_id, "order") '
+                'VALUES ((SELECT id FROM "group" WHERE name = %s), %s, -1);',
+                conn=conn, params=(group, dash_id,), with_fetch=False
+            )
+        elif isinstance(group, dict):
+            self.execute_query(
+                'INSERT INTO dash_group (group_id, dash_id, "order") '
+                'VALUES ((SELECT id FROM "group" WHERE name = %s), %s, %s);',
+                conn=conn, params=(group['name'], dash_id, group['order']), with_fetch=False
+            )
+
     def add_dash(self, *, name, body, groups=None):
         dash_id = self.check_dash_exists(dash_name=name)
         if dash_id:
@@ -698,9 +726,8 @@ class PostgresConnector(PGConnector):
                                       conn=conn, params=(name, body,), as_obj=True)
             if isinstance(groups, list):
                 for group in groups:
-                    self.execute_query('INSERT INTO dash_group (group_id, dash_id) '
-                                       'VALUES ((SELECT id FROM "group" WHERE name = %s), %s);',
-                                       conn=conn, params=(group, dash.id,), with_fetch=False)
+                    self._add_dash_to_group(dash.id, group, conn)
+
         return dash.id, dash.modified
 
     def update_dash(self, *, dash_id, name, body, groups=None):
@@ -718,22 +745,38 @@ class PostgresConnector(PGConnector):
             if body:
                 self.execute_query("UPDATE dash SET body = %s WHERE id = %s;",
                                    conn=conn, params=(body, dash_id), with_fetch=False)
-
         if isinstance(groups, list):
             current_groups = self.execute_query('SELECT name FROM "group" WHERE id IN '
                                                 '(SELECT group_id FROM dash_group WHERE dash_id = %s);',
                                                 params=(dash_id,), fetchall=True)
             current_groups = flat_to_set(current_groups)
-            target_groups = set(groups)
+
+            target_groups = set()
+            if len(groups) > 0:
+                if isinstance(groups[0], str):
+                    target_groups = set(groups)
+                elif isinstance(groups[0], dict):
+                    target_groups = {group['name'] for group in groups}
 
             groups_for_add = target_groups - current_groups
             groups_for_delete = tuple(current_groups - target_groups)
 
             with self.transaction('update_dash_groups') as conn:
-                for group in groups_for_add:
-                    self.execute_query('INSERT INTO dash_group (group_id, dash_id) '
-                                       'VALUES ((SELECT id FROM "group" WHERE name = %s), %s);',
-                                       conn=conn, params=(group, dash_id,), with_fetch=False)
+                for group in groups:
+                    group_name = group['name'] if isinstance(group, dict) else group
+                    if group_name in groups_for_add:
+                        self._add_dash_to_group(dash_id, group, conn)
+                    if group_name in current_groups and isinstance(group, dict):
+                        self.execute_query(
+                            """
+                            UPDATE dash_group SET "order" = %s
+                            WHERE dash_id=%s
+                            AND group_id=(SELECT id FROM "group" WHERE name = %s);
+                            """,
+                            conn=conn, params=(group['order'], dash_id, group_name), with_fetch=False
+                        )
+
+
                 if groups_for_delete:
                     self.execute_query('DELETE FROM dash_group  WHERE group_id IN '
                                        '(SELECT id FROM "group" WHERE name IN %s) AND dash_id = %s;',