+++ sql/update_to_1.15.0.sql
@@ -1,0 +1,3 @@
+ALTER TABLE dash_group
+ADD COLUMN "order" INT NOT NULL DEFAULT -1;
+