+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.15.0] - 2022-09-01
+### Fixed
+ - Wrong timestamps in Timelines
+ - Logging with osr_hid
+ - Deleting one session
+
+### Added
+- Dashboard groups order while saving
+
 ## [1.14.6] - 2022-08-05
 ### Fixed
  - Tests after altered CachesDL table with new "hash" column.